// src/types/data-type.ts
export interface DataProp {
  id: string | number;
  image: string;
  details: string;
  miles: string;
  price: string;
  deposite: string;
  instock: string;
  warranty: string;
  detailes1: string;
  detailes2: string;
  detailes3: string;
  detailes4: string;
  detailes5: string;
  detailes6: string;
  detailes7: string;
  detailes8: string;
  detailes9: string;
  detailes10: string;
}